+++ db/migrations/20161115084604_add_name_and_email.rb
@@ -1,0 +1,9 @@
+Hanami::Model.migration do
+  change do
+    alter_table(:hosts) do
+      add_column :email, String
+      add_column :name,  String
+      add_column :phone, String
+    end
+  end
+end