+++ lib/concierge/flows/host_creation.rb
@@ -157,6 +157,7 @@ module Concierge::Flows
   # background workers. Hosts belong to a supplier and other related
   # attributes.
   class HostCreation
+    include Concierge::JSON
     include Hanami::Validations
 
     IDLE = "idle"
@@ -164,8 +165,12 @@ module Concierge::Flows
     attribute :supplier,       presence: true
     attribute :identifier,     presence: true
     attribute :username,       presence: true
-    attribute :access_token,   presence: true
     attribute :fee_percentage, presence: true
+    attribute :access_token
+    attribute :name
+    attribute :email
+    attribute :phone
+    attribute :payment_terms
 
     attr_reader :config_path
 
@@ -182,6 +187,11 @@ module Concierge::Flows
     # for the supplier the host belongs to.
     def perform
       if valid?
+        if self.access_token.to_s.empty?
+          result = create_roomorama_user
+          return result unless result.success?
+          self.access_token = result.value
+        end
         transaction do
           host = create_host
           workers_definition = find_workers_definition
@@ -216,6 +226,21 @@ module Concierge::Flows
 
     private
 
+    def create_roomorama_user
+      client = Roomorama::Client.new(ENV["CONCIERGE_CREATE_HOST_TOKEN"])
+      creation = Roomorama::Client::Operations::CreateHost.new(supplier, username, name, email, phone, payment_terms)
+      Concierge.context = Concierge::Context.new(type: "host_creation")
+      post_result = client.perform(creation)
+      return post_result unless post_result.success?
+      decode_result = json_decode(post_result.value.body)
+
+      if decode_result.success?
+        return Result.new(decode_result.value["access_token"])
+      else
+        return decode_result
+      end
+    end
+
     def create_host
       existing = HostRepository.from_supplier(supplier).identified_by(identifier).first
       host = existing || Host.new(supplier_id: supplier.id, identifier: identifier)
@@ -223,6 +248,10 @@ module Concierge::Flows
       host.username       = username
       host.access_token   = access_token
       host.fee_percentage = fee_percentage.to_f
+      host.email          = email
+      host.name           = name
+      host.phone          = phone
+      host.payment_terms  = payment_terms
       HostRepository.persist(host)
     end